requirements.txt
certifi==2019.3.9
dj-database-url==0.5.0
Django==2.2.2
-e git://github.com/lepri/django-heroku.git#egg=django-heroku
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
